refactor: introduce NetworkNodeServicesBuilder for improved service configuration
